Basic world object
Just the class for the time being. Most of the changes here involve a
refactoring of the renderer to move visual properties out of the game
logic. Tile size, and the drawing positions that depend on that, are
isolated to the renderer. The game itself is only aware of tile indices.